[Design] location -> photoLocation 수정
